+++ lib/aliyun_oss/bucket/acl.ex
@@ -3,7 +3,7 @@ defmodule Aliyun.Oss.Bucket.ACL do
   Bucket operations - ACL.
   """
 
-  import Aliyun.Oss.Bucket, only: [get_bucket: 4, put_bucket: 5]
+  import Aliyun.Oss.Bucket, only: [get_bucket: 3, put_bucket: 4]
   alias Aliyun.Oss.Config
   alias Aliyun.Oss.Client.{Response, Error}
 
@@ -32,7 +32,7 @@ defmodule Aliyun.Oss.Bucket.ACL do
   """
   @spec get(Config.t(), String.t()) :: {:error, error()} | {:ok, Response.t()}
   def get(config, bucket) do
-    get_bucket(config, bucket, %{}, %{"acl" => nil})
+    get_bucket(config, bucket, %{"acl" => nil})
   end
 
   @doc """
@@ -73,6 +73,6 @@ defmodule Aliyun.Oss.Bucket.ACL do
   @spec put(Config.t(), String.t(), String.t()) ::
           {:error, error()} | {:ok, Aliyun.Oss.Client.Response.t()}
   def put(config, bucket, acl) do
-    put_bucket(config, bucket, %{"x-oss-acl" => acl}, %{"acl" => acl}, "")
+    put_bucket(config, bucket, %{"x-oss-acl" => acl, "acl" => acl}, "")
   end
 end