+++ lib/aliyun_oss/bucket/lifecycle.ex
@@ -3,7 +3,7 @@ defmodule Aliyun.Oss.Bucket.Lifecycle do
   Bucket operations - Lifecycle.
   """
 
-  import Aliyun.Oss.Bucket, only: [get_bucket: 4, put_bucket: 5, delete_bucket: 3]
+  import Aliyun.Oss.Bucket, only: [get_bucket: 3, put_bucket: 4, delete_bucket: 3]
   alias Aliyun.Oss.Config
   alias Aliyun.Oss.Client.{Response, Error}
 
@@ -66,7 +66,7 @@ defmodule Aliyun.Oss.Bucket.Lifecycle do
   end
 
   def put(config, bucket, config) do
-    put_bucket(config, bucket, %{}, %{"lifecycle" => nil}, config)
+    put_bucket(config, bucket, %{"lifecycle" => nil}, config)
   end
 
   @doc """
@@ -97,7 +97,7 @@ defmodule Aliyun.Oss.Bucket.Lifecycle do
   """
   @spec get(Config.t(), String.t()) :: {:error, error()} | {:ok, Response.t()}
   def get(config, bucket) do
-    get_bucket(config, bucket, %{}, %{"lifecycle" => nil})
+    get_bucket(config, bucket, %{"lifecycle" => nil})
   end
 
   @doc """