+++ lib/aliyun_oss/bucket/request_payment.ex
@@ -3,7 +3,7 @@ defmodule Aliyun.Oss.Bucket.RequestPayment do
   Bucket operations - RequestPayment.
   """
 
-  import Aliyun.Oss.Bucket, only: [get_bucket: 4, put_bucket: 5]
+  import Aliyun.Oss.Bucket, only: [get_bucket: 3, put_bucket: 4]
   alias Aliyun.Oss.Config
   alias Aliyun.Oss.Client.{Response, Error}
 
@@ -29,7 +29,7 @@ defmodule Aliyun.Oss.Bucket.RequestPayment do
   """
   @spec get(Config.t(), String.t()) :: {:error, error()} | {:ok, Response.t()}
   def get(config, bucket) do
-    get_bucket(config, bucket, %{}, %{"requestPayment" => nil})
+    get_bucket(config, bucket, %{"requestPayment" => nil})
   end
 
   @doc """
@@ -58,6 +58,6 @@ defmodule Aliyun.Oss.Bucket.RequestPayment do
   @spec put(Config.t(), String.t(), String.t()) :: {:error, error()} | {:ok, Response.t()}
   def put(config, bucket, payer) do
     xml_body = EEx.eval_string(@body_tmpl, payer: payer)
-    put_bucket(config, bucket, %{}, %{"requestPayment" => nil}, xml_body)
+    put_bucket(config, bucket, %{"requestPayment" => nil}, xml_body)
   end
 end